Updated for cart changes to reflect

+++ src/main/java/com/cosmos/model/Product.java
@@ -27,10 +27,10 @@ public class Product {
 	private String productBrand;
 	private String productStore;
 	private String productUnit;
-	private int productAmount;
-	private int productMRP;
-	private int productRealPrice;
-	private int productSellingPrice;
+	private Float productAmount;
+	private Float productMRP;
+	private Float productRealPrice;
+	private Float productSellingPrice;
 	private boolean active;
 	private LocalDate now;
 	private boolean addedToCart;
